+++ package.json
@@ -13,6 +13,7 @@
     "ansis": "^3.2.0",
     "change-case": "^5.4.4",
     "cli-spinners": "^2",
+    "figures": "^6.1.0",
     "ink": "^5.0.1",
     "is-wsl": "^3.1.0",
     "open": "^10.1.0",
