+++ src/components/divider.tsx
@@ -1,0 +1,54 @@
+/*
+ * Copyright (c) 2024, salesforce.com, inc.
+ * All rights reserved.
+ * Licensed under the BSD 3-Clause license.
+ * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
+ */
+import { Box, Text } from 'ink';
+import React from 'react';
+
+const getSideDividerWidth = (width: number, titleWidth: number): number => (width - titleWidth) / 2;
+const getNumberOfCharsPerWidth = (char: string, width: number): number => width / char.length;
+
+const PAD = ' ';
+
+export function Divider({
+  title = '',
+  width = 50,
+  padding = 1,
+  titlePadding = 1,
+  titleColor = 'white',
+  dividerChar = '─',
+  dividerColor = 'dim',
+}: {
+  readonly title?: string;
+  readonly width?: number | 'full';
+  readonly padding?: number;
+  readonly titleColor?: string;
+  readonly titlePadding?: number;
+  readonly dividerChar?: string;
+  readonly dividerColor?: string;
+}): React.ReactNode {
+  const titleString = title ? `${PAD.repeat(titlePadding) + title + PAD.repeat(titlePadding)}` : '';
+  const titleWidth = titleString.length;
+  const terminalWidth = process.stdout.columns ?? 80;
+  const widthToUse = width === 'full' ? terminalWidth - titlePadding : width > terminalWidth ? terminalWidth : width;
+
+  const dividerWidth = getSideDividerWidth(widthToUse, titleWidth);
+  const numberOfCharsPerSide = getNumberOfCharsPerWidth(dividerChar, dividerWidth);
+  const dividerSideString = dividerChar.repeat(numberOfCharsPerSide);
+
+  const paddingString = PAD.repeat(padding);
+
+  return (
+    <Box flexDirection="row">
+      <Text>
+        {paddingString}
+        <Text color={dividerColor}>{dividerSideString}</Text>
+        <Text color={titleColor}>{titleString}</Text>
+        <Text color={dividerColor}>{dividerSideString}</Text>
+        {paddingString}
+      </Text>
+    </Box>
+  );
+}
